+++ test/test_pycurves.py
@@ -1,8 +1,6 @@
 import matplotlib.pyplot as plt
 import openpile.utils.py_curves as py
 import pytest 
+import math as m
 
-
-# @pytest.xfail
-# def test_api_sand():
-#     assert 2 == 3
+import openpile.utils.py_curves as py